# api/openapi.yaml
openapi: 3.0.3
info:
  title: iExec market API
  description: >-
    iExec market API OpenAPI specification
  version: 6.3.0
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: https://api.market.iex.ec
  - url: http://localhost:3000
paths:
...
  /challenge:
    get:
      description: get a challenge to sign for authenticated operations
      parameters:
        - $ref: '#/components/parameters/ChainId'
        - name: address
          in: query
          description: ethereum address of the wallet to authenticate
          required: true
          schema:
            type: string
            example: '0xF048eF3d7E3B33A465E0599E641BB29421f7Df92'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - $ref: '#/components/schemas/EIP712'
  /categories/{catId}:
    get:
      description: get a protocol category by id
      parameters:
        - $ref: '#/components/parameters/ChainId'
        - name: catId
          in: path
          description: category id
          required: true
          schema:
            type: number
            example: 0
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - $ref: '#/components/schemas/Category'
  /categories:
    get:
      description: get the protocol categories
      parameters:
        - $ref: '#/components/parameters/ChainId'
        - name: minWorkClockTimeRef
          in: query
          description: smallest workClockTimeRef to filter results on
          schema:
            type: number
        - name: maxWorkClockTimeRef
          in: query
          description: biggest workClockTimeRef to filter results on
          schema:
            type: number
        - $ref: '#/components/parameters/PageIndex'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Page'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - $ref: '#/components/schemas/PaginableResponse'
                  - type: object
                    properties:
                      categories:
                        type: array
                        items:
                          $ref: '#/components/schemas/Category'
  /deals/{dealId}:
    get:
      description: get a deal by id
      parameters:
        - $ref: '#/components/parameters/ChainId'
        - name: dealId
          in: path
          description: deal id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - $ref: '#/components/schemas/Deal'
  /deals:
    get:
      description: get the protocol deals
      parameters:
        - $ref: '#/components/parameters/ChainId'
        - name: category
          in: query
          description: category id to filter results on
          schema:
            type: number
        - name: requester
          in: query
          description: requester address to filter results on
          schema:
            type: string
        - name: beneficiary
          in: query
          description: beneficiary address to filter results on
          schema:
            type: string
        - name: app
          in: query
          description: app address to filter results on
          schema:
            type: string
        - name: dataset
          in: query
          description: dataset address to filter results on
          schema:
            type: string
        - name: workerpool
          in: query
          description: workerpool address to filter results on
          schema:
            type: string
        - name: appOwner
          in: query
          description: app owner address to filter results on
          schema:
            type: string
        - name: datasetOwner
          in: query
          description: dataset owner address to filter results on
          schema:
            type: string
        - name: workerpoolOwner
          in: query
          description: workerpool owner address to filter results on
          schema:
            type: string
        - name: apporderHash
          in: query
          description: apporder hash to filter results on
          schema:
            type: string
        - name: datasetorderHash
          in: query
          description: datasetorder hash to filter results on
          schema:
            type: string
        - name: workerpoolorderHash
          in: query
          description: workerpoolorder hash to filter results on
          schema:
            type: string
        - name: requestHash
          in: query
          description: requestorder hash to filter results on
          schema:
            type: string
        - $ref: '#/components/parameters/PageIndex'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Page'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - $ref: '#/components/schemas/PaginableResponse'
                  - type: object
                    properties:
                      deals:
                        type: array
                        items:
                          $ref: '#/components/schemas/Deal'
  /apporders/{orderHash}:
    get:
      description: get a apporder by hash
      parameters:
        - $ref: '#/components/parameters/ChainId'
        - $ref: '#/components/parameters/OrderHash'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - $ref: '#/components/schemas/AppOrderWithMetadata'
  /apporders:
    get:
      description: browse the apporders market
      parameters:
        - $ref: '#/components/parameters/ChainId'
        - name: app
          in: query
          description: app address to filter results on (accepts address or "any")
          schema:
            type: string
            example: 'any'
        - name: isAppStrict
          in: query
          description: exclude orders with permissions allowing "any" app (default false, ignored if `app` is not specified)
          schema:
            type: boolean
        - name: appOwner
          in: query
          description: app owner address to filter results on
          schema:
            type: string
        - name: minTag
          in: query
          description: minimum tag to filter results on
          schema:
            type: string
        - name: maxTag
          in: query
          description: maximum tag to filter results on
          schema:
            type: string
        - name: minVolume
          in: query
          description: minimum volume to filter results on
          schema:
            type: number
        - name: dataset
          in: query
          description: datasetrestrict address to include in results
          schema:
            type: string
        - name: isDatasetStrict
          in: query
          description: exclude orders with permissions allowing "any" dataset (default false, ignored if `dataset` is not specified)
          schema:
            type: boolean
        - name: workerpool
          in: query
          description: workerpoolrestrict address to include in results
          schema:
            type: string
        - name: isWorkerpoolStrict
          in: query
          description: exclude orders with permissions allowing "any" workerpool (default false, ignored if `workerpool` is not specified)
          schema:
            type: boolean
        - name: requester
          in: query
          description: requesterrestrict address to include in results
          schema:
            type: string
        - name: isRequesterStrict
          in: query
          description: exclude orders with permissions allowing "any" requester (default false, ignored if `requester` is not specified)
          schema:
            type: boolean
        - $ref: '#/components/parameters/PageIndex'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Page'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - $ref: '#/components/schemas/PaginableResponse'
                  - type: object
                    properties:
                      orders:
                        type: array
                        items:
                          $ref: '#/components/schemas/AppOrderWithMetadata'
    post:
      description: publish an apporder on the market
      parameters:
        - $ref: '#/components/parameters/ChainId'
        - $ref: '#/components/parameters/Authorization'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                order:
                  $ref: '#/components/schemas/AppOrder'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      published:
                        $ref: '#/components/schemas/AppOrderWithMetadata'
    put:
      description: unpublish apporders
      parameters:
        - $ref: '#/components/parameters/ChainId'
        - $ref: '#/components/parameters/Authorization'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                target:
                  type: string
                  enum:
                    - unpublish_orderHash
                    - unpublish_last
                    - unpublish_all
                orderHash:
                  type: string
                app:
                  type: string

      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      unpublished:
                        type: array
                        items:
                          type: string
  /datasetorders/{orderHash}:
    get:
      description: get a datasetorder by hash
      parameters:
        - $ref: '#/components/parameters/ChainId'
        - $ref: '#/components/parameters/OrderHash'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      published:
                        $ref: '#/components/schemas/DatasetOrderWithMetadata'
...
  /workerpoolorders/{orderHash}:
    get:
      description: get a workerpoolorder by hash
      parameters:
        - $ref: '#/components/parameters/ChainId'
        - $ref: '#/components/parameters/OrderHash'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - $ref: '#/components/schemas/WorkerpoolOrderWithMetadata'
...
  /requestorders/{orderHash}:
    get:
      description: get a requestorder by hash
      parameters:
        - $ref: '#/components/parameters/ChainId'
        - $ref: '#/components/parameters/OrderHash'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - $ref: '#/components/schemas/RequestOrderWithMetadata'
...
components:
  parameters:
    ChainId:
      name: chainId
      in: query
      description: chain id
      required: true
      schema:
        type: number
        example: 134
...
    OrderHash:
      name: orderHash
      in: path
      required: true
      description: order hash
      schema:
        type: string
    Page:
      name: page
      in: query
      description: deprecated pagination param
      deprecated: true
      schema:
        type: number
    PageIndex:
      name: pageIndex
      in: query
      schema:
        type: number
    PageSize:
      name: pageSize
      in: query
      schema:
        type: number
  schemas:
    ApiResponse:
      properties:
        ok:
          type: boolean
...
    PaginableResponse:
      properties:
        count:
          type: number
...
    EIP712:
      properties:
        ok:
          type: boolean
        data:
          type: object
          properties:
            types:
              type: object
              properties:
                EIP712Domain:
                  type: array
                  items:
                    type: object
                    properties:
                      name:
                        type: string
                      type:
                        type: string
                Challenge:
                  type: array
                  items:
                    type: object
                    properties:
                      name:
                        type: string
                      type:
                        type: string
        domain:
          type: object
          properties:
            name:
              type: string
            version:
              type: string
            chainId:
              type: string
        primaryType:
          type: string
        message:
          type: object
          properties:
            challenge:
              type: string
    Category:
      properties:
        catid:
          type: number
        chainId:
          type: string
        name:
          type: string
        description:
          type: string
        workClockTimeRef:
          type: number
        transactionHash:
          type: string
        blockNumber:
          type: number
        blockTimestamp:
          type: string
...